+++ packages/json-rpc/polkadot-sdk-compat/src/parsed-enhancers/index.ts
@@ -1,3 +1,4 @@
+export { fixUnorderedBlocks } from "./fix-unordered-blocks"
 export { fixUnorderedEvents } from "./fix-unordered-events"
 export { patchChainHeadEvents } from "./patch-chainhead-events"
 export { unpinHash } from "./unpin-hash"